+++ pyccel/ast/core.py
@@ -116,7 +116,7 @@ class AsName(PyccelAstNode):
     ----------
     obj : PyccelAstNode or PyccelAstNodeType
         The variable, function, or module being renamed.
-    target : str
+    local_alias : str
         Name of variable or function in this context.
 
     Examples
@@ -129,16 +129,16 @@ class AsName(PyccelAstNode):
     >>> AsName(NumpyFull, 'fill_func')
     full as fill_func
     """
-    __slots__ = ('_obj', '_target')
+    __slots__ = ('_obj', '_local_alias')
     _attribute_nodes = ()
 
-    def __init__(self, obj, target):
+    def __init__(self, obj, local_alias):
         if pyccel_stage != "syntactic":
             assert (isinstance(obj, PyccelAstNode) and \
                     not isinstance(obj, PyccelSymbol)) or \
                    (isinstance(obj, type) and issubclass(obj, PyccelAstNode))
         self._obj = obj
-        self._target = target
+        self._local_alias = local_alias
         super().__init__()
 
     @property
@@ -152,10 +152,13 @@ class AsName(PyccelAstNode):
             return obj.name
 
     @property
-    def target(self):
-        """ The target name of the object
+    def local_alias(self):
         """
-        return self._target
+        The local_alias name of the object.
+
+        The name used to identify the object in the local scope.
+        """
+        return self._local_alias
 
     @property
     def object(self):
@@ -164,13 +167,13 @@ class AsName(PyccelAstNode):
         return self._obj
 
     def __repr__(self):
-        return f'{self.object} as {self.target}'
+        return f'{self.object} as {self.local_alias}'
 
     def __eq__(self, string):
         if isinstance(string, str):
-            return string == self.target
+            return string == self.local_alias
         elif isinstance(string, AsName):
-            return string.target == self.target
+            return string.local_alias == self.local_alias
         else:
             return self is string
 
@@ -178,7 +181,7 @@ class AsName(PyccelAstNode):
         return not self == string
 
     def __hash__(self):
-        return hash(self.target)
+        return hash(self.local_alias)
 
 
 class Duplicate(TypedAstNode):
@@ -646,10 +649,10 @@ class CodeBlock(PyccelAstNode):
 
 class AliasAssign(PyccelAstNode):
     """
-    Representing assignment of an alias to its target.
+    Representing assignment of an alias to its local_alias.
 
     Represents aliasing for code generation. An alias is any statement of the
-    form `lhs := rhs` where lhs is a pointer and rhs is a target. In other words
+    form `lhs := rhs` where lhs is a pointer and rhs is a local_alias. In other words
     the contents of `lhs` will change if the contents of `rhs` are modfied.
 
     Parameters
@@ -664,7 +667,7 @@ class AliasAssign(PyccelAstNode):
            Variable.
 
     rhs : PyccelSymbol | Variable, IndexedElement
-        The target of the assignment. A PyccelSymbol in the syntactic stage,
+        The local_alias of the assignment. A PyccelSymbol in the syntactic stage,
         a Variable or a Slice of an array in the semantic stage.
 
     Examples
@@ -3838,9 +3841,25 @@ class Import(PyccelAstNode):
             self._target[new_target] = None
 
     def find_module_target(self, new_target):
+        """
+        Find the specified target amongst the targets of the Import.
+
+        Find the specified target amongst the targets of the Import.
+
+        Parameters
+        ----------
+        new_target : str
+            The name of the target that has been imported.
+
+        Returns
+        -------
+        str
+            The name of the target in the local scope or None if the
+            target is not found.
+        """
         for t in self._target:
             if isinstance(t, AsName) and new_target == t.name:
-                return t.target
+                return t.local_alias
             elif new_target == t:
                 return t
         return None